+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] - 2024-01-08
+
+This version adds a roughed-in sandbox for scheme code and two server
+configs to support it: `scruple.call_time_limit` and
+`scruple.call_allocation_limit`.
+
 ## [0.16.0] - 2024-01-06
 
 This version adds support for "do" statements.